ci: Fix matrix to use testgui flag properly
This flag is there to get CI to only run either GUI or CLI tests, unless
we are publishing a release, in order to cut down on interation time.
When the YAML file was split into two in #1559, there was a mistake
where this matrix flag wasn't ported over properly to an input, causing
it to always evaluate to false.