Revamp README.md
The existing MacVim README has been barebones without much useful
information for a while now. Revamp it to be more standard and include a
screenshot. More may be added later when we add more overall
documentation and can be linked from the README.

While we are at it, upload light/dark versions of the screenshot to fit
the user's preferences.