CI: Always upload package cache even if overall build fails
When we don't trigger CI for more than 1 week, or if the
gettext/libsodium packages were updated, CI needs to re-generate the
artifacts and cache them. However, if the overall build fails later due
to unrelated test failures, previously it would stop the cache from
being uploaded at all, making iterating on fixing the test failure quite
painful as we need to rebuild the artifact every run until it passes.
This change now makes it so we always upload the cache as long as the
package was built correctly, so that CI runs after the first failure
will be much smoother.