CI: Fix failure to install Homebrew package in universal-package
The universal-package action tries to build a custom formula and
installs it by calling `brew install` on a raw .rb file. This recently
started to fail as Homebrew added a change to consider this use case as
an error. Change the action to use the officially sanctioned method
which is to create a tap and then place the formula in said tap.