Allow underscores in guifont PostScript names
MacVim guifont's can be specified by either full PostScript names or
family names. Since family names can contain spaces, MacVim allows using
underscores to serve as substitute to make it easier for the user to
type, and this is similar to how Win32 gVim works.

However, some fonts use underscores in their full PostScript name. For
example, Cascadia Code has names like "CascadiaCode-Regular_Light". The
previous substitution code always replaced underscores with spaces which
caused them to fail to load. Fix the code to only do the
underscore-space substitution if the font could not be loaded. Also, fix
documentation to be clearer in what MacVim expects as guifont's input.

In retrospect, implicit substitution like this was probably a bad idea
when the user could just escape with spaces, but given this has been the
behavior for a long time we should just keep it to avoid breaking
backwards compatibility.

Fix #1208