ci: Add macos-26, and swap out macos-13
Don't use macos-26 / Xcode 26 for making release builds yet. Doing so
will require updating the OS deployment target and move legacy builds to
include 10.13-10.15, and MacVim also doesn't work perfectly when
compiled under the macOS 26 SDK right now.

With this change we also no longer have any x86 runners in the mix.
While GitHub Actions provides a macos-15-intel runner, it will be
deprecate in near future as well, so it's not really worth setting it
up. If we care to test MacVim on x86 machines we may have to find
alternative solutions, but for now it should be ok as we haven't seen
any issues with universal builds for a long time.