ci: Remove macos-13 runner
macOS-13 is retired from GitHub hosted runners, so we need to remove it.

With this change we also no longer have any x86 runners in the mix.
While GitHub Actions provides a macos-15-intel runner, it will be
deprecate in near future as well, so it's not really worth setting it
up. If we care to test MacVim on x86 machines we may have to find
alternative solutions, but for now it should be ok as we haven't seen
any issues with universal builds for a long time.